Move ServerOptionsBuilderOptions into test assembly
Remove static Build() method from ServerOptionsBuilder
Make the constructor public
Make validation methods protected virtual 
Override validation methods from ServerOptionsBuilderFake
