+++ src/sestest/ServerOptionBuilder.cs
@@ -18,36 +18,14 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement
         // Reference to a store in which we can search for certificates
         private readonly CertificateStore _certificateStore = new CertificateStore();
 
-        // Options used to configure validation
-        private readonly ServerOptionBuilderOptions _options;
-
-        /// <summary>
-        /// Build an instance of <see cref="ServerOptions"/> from the information supplied on the 
-        /// command line by the user
-        /// </summary>
-        /// <param name="commandLine">Command line parameters supplied by the user.</param>
-        /// <param name="options">Options for configuring validation.</param>
-        /// <returns>Either a usable (and completely valid) <see cref="ServerOptions"/> or a set 
-        /// of errors.</returns>
-        public static Errorable<ServerOptions> Build(
-            ServerCommandLine commandLine,
-            ServerOptionBuilderOptions options = ServerOptionBuilderOptions.None)
-        {
-            var builder = new ServerOptionBuilder(commandLine, options);
-            return builder.Build();
-        }
-
         /// <summary>
         /// Initializes a new instance of the <see cref="ServerOptionBuilder"/> class
         /// </summary>
         /// <param name="commandLine">Options provided on the command line.</param>
-        /// <param name="options">Options for configuring validation.</param>
-        private ServerOptionBuilder(
-            ServerCommandLine commandLine,
-            ServerOptionBuilderOptions options)
+        public ServerOptionBuilder(
+            ServerCommandLine commandLine)
         {
             _commandLine = commandLine;
-            _options = options;
         }
 
         /// <summary>
@@ -56,7 +34,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement
         /// </summary>
         /// <returns>Either a usable (and completely valid) <see cref="ServerOptions"/> or a set 
         /// of errors.</returns>
-        private Errorable<ServerOptions> Build()
+        public Errorable<ServerOptions> Build()
         {
             var result = Errorable.Success(new ServerOptions())
                 .Configure(ServerUrl(), (opt, url) => opt.WithServerUrl(url))
@@ -75,15 +53,10 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement
         /// </summary>
         /// <returns>An <see cref="Errorable{Uri}"/> containing either the URL to use or any 
         /// relevant errors.</returns>
-        private Errorable<Uri> ServerUrl()
+        protected virtual Errorable<Uri> ServerUrl()
         {
             if (string.IsNullOrWhiteSpace(_commandLine.ServerUrl))
             {
-                if (_options.HasFlag(ServerOptionBuilderOptions.ServerUrlOptional))
-                {
-                    return Errorable.Success(new Uri("http://test"));
-                }
-
                 return Errorable.Failure<Uri>("No server endpoint URL specified.");
             }
 
@@ -107,15 +80,10 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement
         /// Find the certificate to use for HTTPS connections
         /// </summary>
         /// <returns>Certificate, if found; error details otherwise.</returns>
-        private Errorable<X509Certificate2> ConnectionCertificate()
+        protected virtual Errorable<X509Certificate2> ConnectionCertificate()
         {
             if (string.IsNullOrEmpty(_commandLine.ConnectionCertificateThumbprint))
             {
-                if (_options.HasFlag(ServerOptionBuilderOptions.ConnectionThumbprintOptional))
-                {
-                    return Errorable.Success<X509Certificate2>(null);
-                }
-
                 return Errorable.Failure<X509Certificate2>("A connection thumbprint is required.");
             }
 
@@ -205,13 +173,5 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement
             var certificateThumbprint = new CertificateThumbprint(thumbprint);
             return _certificateStore.FindByThumbprint(purpose, certificateThumbprint);
         }
-    }
-
-    [Flags]
-    public enum ServerOptionBuilderOptions
-    {
-        None,
-        ServerUrlOptional,
-        ConnectionThumbprintOptional
     }
 }
