+++ tests/Microsoft.Azure.Batch.SoftwareEntitlement.Tests/ServerOptionBuilderTests.cs
@@ -25,7 +25,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             [Fact]
             public void Build_WithEmptyServerUrl_DoesNotReturnValue()
             {
-                var options = ServerOptionBuilder.Build(_commandLine);
+                var options = new ServerOptionBuilder(_commandLine).Build();
                 options.HasValue.Should().BeFalse();
             }
 
@@ -33,7 +33,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void Build_WithEmptyServerUrl_HasErrorForServerUrl()
             {
                 _commandLine.ServerUrl = string.Empty;
-                var options = ServerOptionBuilder.Build(_commandLine);
+                var options = new ServerOptionBuilder(_commandLine).Build();
                 options.Errors.Should().Contain(e => e.Contains("server endpoint URL"));
             }
 
@@ -41,7 +41,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void Build_WithHttpServerUrl_HasErrorForServerUrl()
             {
                 _commandLine.ServerUrl = "http://www.example.com";
-                var options = ServerOptionBuilder.Build(_commandLine);
+                var options = new ServerOptionBuilder(_commandLine).Build();
                 options.Errors.Should().Contain(e => e.Contains("Server endpoint URL"));
             }
 
@@ -49,7 +49,9 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void WithValidServerUrl_ConfigureServerUrl()
             {
                 _commandLine.ServerUrl = "https://example.com/";
-                var options = ServerOptionBuilder.Build(_commandLine, _permissiveOptions);
+                var options = new ServerOptionBuilderFake(
+                    _commandLine, ServerOptionBuilderOptions.ConnectionThumbprintOptional)
+                    .Build();
                 options.Value.ServerUrl.ToString().Should().Be(_commandLine.ServerUrl);
             }
         }
@@ -60,7 +62,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void Build_WithNoConnectionThumbprint_DoesNotReturnValue()
             {
                 _commandLine.ConnectionCertificateThumbprint = string.Empty;
-                var options = ServerOptionBuilder.Build(_commandLine);
+                var options = new ServerOptionBuilder(_commandLine).Build();
                 options.HasValue.Should().BeFalse();
             }
 
@@ -68,7 +70,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void Build_WithNoConnectionThumbprint_HasErrorForConnection()
             {
                 _commandLine.ConnectionCertificateThumbprint = string.Empty;
-                var options = ServerOptionBuilder.Build(_commandLine);
+                var options = new ServerOptionBuilder(_commandLine).Build();
                 options.Errors.Should().Contain(e => e.Contains("connection"));
             }
 
@@ -76,7 +78,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void Build_WithUnknownConnectionThumbprint_HasErrorForConnection()
             {
                 _commandLine.ConnectionCertificateThumbprint = _thumbprint;
-                var options = ServerOptionBuilder.Build(_commandLine);
+                var options = new ServerOptionBuilder(_commandLine).Build();
                 options.Errors.Should().Contain(e => e.Contains("connection"));
             }
         }
@@ -87,7 +89,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void Build_WithEmptyAudience_HasDefaultValueForAudience()
             {
                 _commandLine.Audience = string.Empty;
-                var options = ServerOptionBuilder.Build(_commandLine, _permissiveOptions);
+                var options = new ServerOptionBuilderFake(_commandLine, _permissiveOptions).Build();
                 options.Value.Audience.Should().NotBeNullOrEmpty();
             }
         }
@@ -98,7 +100,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void Build_WithEmptyIssuer_HasDefaultValueForIssuer()
             {
                 _commandLine.Issuer = string.Empty;
-                var options = ServerOptionBuilder.Build(_commandLine, _permissiveOptions);
+                var options = new ServerOptionBuilderFake(_commandLine, _permissiveOptions).Build();
                 options.Value.Issuer.Should().NotBeNullOrEmpty();
             }
         }
@@ -109,7 +111,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void Build_WithoutExitAfterRequest_ShouldHaveDefault()
             {
                 _commandLine.ExitAfterRequest = false;
-                var options = ServerOptionBuilder.Build(_commandLine, _permissiveOptions);
+                var options = new ServerOptionBuilderFake(_commandLine, _permissiveOptions).Build();
                 options.Value.ExitAfterRequest.Should().BeFalse();
             }
 
@@ -117,7 +119,7 @@ namespace Microsoft.Azure.Batch.SoftwareEntitlement.Tests
             public void Build_WithExitAfterRequest_ConfiguresValue()
             {
                 _commandLine.ExitAfterRequest = true;
-                var options = ServerOptionBuilder.Build(_commandLine, _permissiveOptions);
+                var options = new ServerOptionBuilderFake(_commandLine, _permissiveOptions).Build();
                 options.Value.ExitAfterRequest.Should().BeTrue();
             }
         }
